fixing migration KategoriPelajaran

+++ database/migrations/2024_02_19_145539_create_kategori_pelajarans_table.php
@@ -4,20 +4,22 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
+class CreateKategoriPelajaransTable extends Migration
 {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::create('kategori_pelajarans', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('tahun_ajaran_id')->constrained()->onDelete('cascade');
-            $table->string('kode_kategori', 10)->unique();
-            $table->string('nama_kategori', 50)->unique();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('kategori_pelajarans')) {
+            Schema::create('kategori_pelajarans', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('tahun_ajaran_id')->constrained()->onDelete('cascade');
+                $table->string('kode_kategori', 10)->unique();
+                $table->string('nama_kategori', 50)->unique();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
